Implemented basic solver
It recursively tries values for each of the cells.
Once a value has been set, it will see if this implies other cells. For this it checks the following:
- Are there nearby cells that now can only have one possible value?
- Are there constraint groups for which a given value can only have one possible location?
If so, it will set these cells as well. Furthermore, if it detects that a constraint cannot be met, it will immediately backtrack.

The implementation still includes various SerialUSB debugging statements.

Limitations:
- No randomization yet (so not suitable for generating puzzles)
- Can maybe clear numAutoSet array more efficiently?