Further improved solve animation
Color of lights is now not based on single column but on weighted average of two columns. This increased the range of possible colors from four to twelve, thereby making the changes more gradual.

Furthermore, added extra color variation to scrolling text by changing "done" to light green and exclamation mark to yellow.