Added ability to edit puzzle
While editing puzzle, it checks how many solutions there are after every change.
As long as there are multiple solutions, this is indicated by showing all values as green.
If there is now only one solution, it moves to solve mode. Values are fixed and the puzzle looks as it normally does.
If there is no solution, this is indicated by showing all values as red. It is now only possible to clear values until this is resolved.