Added ability to store and load a puzzle
Storing can be done when solving a puzzle, or while creating a puzzle.
After loading a puzzle that is being created, editing can continue. The only exception is when the puzzle was being solved. This is the way to exit the editing mode once creation is done.

Some basic sanity testing has been done, but some more is needed.