Started support for Hyper sudokus
Updating save/load to store Hyper Sudoku in another slot.
Updated drawing to show hyper boxes.

Next to explicit Hyper-box constraint groups, introduced implicit groups. Without these, the solver fails to quickly detect that a given puzzle cannot be solved, causing it to "hang" during its exhaustive search.

With these, there is a need to distinguish allowed from possible values. The user should still be able to enter values that prevent a solution yet obey all the rules of the puzzle.

This still requires further refinement. After the user undoes an impossible value, the bookkeeping for this implicit constraint group is inaccurate. This still prevents the solver from quickly detecting impossibility.

As this is still a WIP, it contains quite some debugging code.