Improved Hyper Sudoku support
Solver now sets the masks for the implicit constraint groups before starting to solve, to reflect that these may not be valid anymore.

To facilitate this, removed cell-specific masks. Instead let cells calculate their mask by referring to the masks for the constraint groups that they are part of.
This required pointer from SudokuCell to Sudoku with some changes to initialisation.

Moved utility bit functions to Utils.